+++ tests/CallbackIteratorTest.php
@@ -1,5 +1,14 @@
 <?php
 
+/*
+ * This file is part of the awesomite/iterators package.
+ *
+ * (c) Bartłomiej Krukowski <bartlomiej@krukowski.me>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
 namespace Awesomite\Iterators;
 
 use Symfony\Component\Process\Process;
@@ -61,8 +70,8 @@ class CallbackIteratorTest extends BaseTestCase
         return array(
             array(array()),
             array(array(1, 2, 3, 4, 5)),
-            array(range(5, 10, .1)),
-            array(array('q', 'w', 'e', 'r', 't', 'y'))
+            array(\range(5, 10, .1)),
+            array(array('q', 'w', 'e', 'r', 't', 'y')),
         );
     }
 
@@ -78,25 +87,52 @@ class CallbackIteratorTest extends BaseTestCase
                 CallbackIterator::stopIterate();
             }
 
-            return mt_rand();
+            return \mt_rand();
         });
 
-        foreach ($iterator as $randValue) {}
-        foreach ($iterator as $randValue) {}
+        foreach ($iterator as $randValue) {
+        }
+        foreach ($iterator as $randValue) {
+        }
     }
 
-    public function testMemory()
+    /**
+     * @dataProvider providerLength
+     *
+     * @param int $length
+     */
+    public function testMemory($length)
     {
-        $arrayProcess = new Process('php memoryUsage.php array', __DIR__);
+        $arrayProcess = new Process('php memoryUsage.php array ' . $length, __DIR__);
         $arrayProcess->run();
-        $callbackProcess = new Process('php memoryUsage.php callback', __DIR__);
+        $callbackProcess = new Process('php memoryUsage.php callback ' . $length, __DIR__);
         $callbackProcess->run();
-        $callbackMemory = (int) $callbackProcess->getOutput();
-        $arrayMemory = (int) $arrayProcess->getOutput();
-        $div = 1024 * 1024;
+        $callbackMemory = (int)$callbackProcess->getOutput();
+        $arrayMemory = (int)$arrayProcess->getOutput();
+        $mb = 1024 * 1024;
         $this->assertTrue(
-            $callbackMemory < $arrayMemory,
-            sprintf('Callback: %0.2f MB, array: %0.2f MB', $callbackMemory / $div, $arrayMemory / $div)
+            $callbackMemory <= ($arrayMemory + $mb),
+            \sprintf('Callback: %0.2f MB, array: %0.2f MB', $callbackMemory / $mb, $arrayMemory / $mb)
+        );
+        \register_shutdown_function(function () use ($callbackMemory, $arrayMemory, $mb, $length) {
+            echo \sprintf(
+                "Length: % 7d, callback: % 6.2f MB, array: % 6.2f MB\n",
+                $length,
+                $callbackMemory / $mb,
+                $arrayMemory / $mb
+            );
+        });
+    }
+
+    public function providerLength()
+    {
+        return array(
+            array(1),
+            array(10),
+            array(100),
+            array(10000),
+            array(100000),
+            array(1000000),
         );
     }
 }