+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 1.0.0
+
+* Support for PHP 7.2
+* Changed exception message in `Awesomite\Iterators\CallbackIterator::rewind`
+* Excluded unused files from git archive in file `.gitattributes`
+
 ## 0.2.0
 
 * Class `Awesomite\Iterators\StopIterateException` has been `@internal`